- Filtering on Scoreboard implemented - Last logics of repeatgoals implemented - Correct logo for HIPER installed

+++ HIPER/EditGoalPage.xaml.cs
@@ -25,16 +25,14 @@ namespace HIPER
             goalCurrentEntry.Text = selectedGoal.CurrentValue;
             privateGoalCheckbox.IsChecked = selectedGoal.PrivateGoal;
 
-            goalNameEntry.IsEnabled = !selectedGoal.Completed;
-            goalDescriptionEntry.IsEnabled = !selectedGoal.Completed;
-            goalDeadlineEntry.IsEnabled = !selectedGoal.Completed;
-            goalTargetEntry.IsEnabled = !selectedGoal.Completed;
-            goalCurrentEntry.IsEnabled = !selectedGoal.Completed;
-            privateGoalCheckbox.IsEnabled = !selectedGoal.Completed;
+            gridTitle.IsEnabled = !(selectedGoal.Completed || selectedGoal.Closed);
+            gridAimHigh.IsEnabled = !(selectedGoal.Completed || selectedGoal.Closed);
+            gridSingle.IsEnabled = !(selectedGoal.Completed || selectedGoal.Closed);
+            gridSingle.IsEnabled = !(selectedGoal.Completed || selectedGoal.Closed);
 
-            deleteGoal.IsVisible = !selectedGoal.Completed;
-            updateGoal.IsVisible = !selectedGoal.Completed;
-            completeGoal.IsVisible = !selectedGoal.Completed;
+            deleteGoal.IsVisible = !(selectedGoal.Completed || selectedGoal.Closed);
+            updateGoal.IsVisible = !(selectedGoal.Completed || selectedGoal.Closed);
+            completeGoal.IsVisible = !(selectedGoal.Completed || selectedGoal.Closed);
 
             repeatableRB1.IsChecked = (selectedGoal.TargetType == 0) ? true : false;
             repeatableRB2.IsChecked = (selectedGoal.TargetType == 1) ? true : false;
@@ -66,8 +64,8 @@ namespace HIPER
 
             bool isGoalNameEmpty = string.IsNullOrEmpty(goalNameEntry.Text);
             bool isGoalDescriptionEmpty = string.IsNullOrEmpty(goalDescriptionEntry.Text);
-            bool isWeeklyCheckedAndEntryFilled = repeatableRB21.IsChecked && (weekdayPicker.SelectedIndex < 0);
-            bool isMonthlyCheckedAndEntryFilled = repeatableRB22.IsChecked && (dayOfMonthPicker.SelectedIndex < 0);
+            bool isWeeklyCheckedAndEntryFilled = repeatableRB2.IsChecked && repeatableRB21.IsChecked && (weekdayPicker.SelectedIndex < 0);
+            bool isMonthlyCheckedAndEntryFilled = repeatableRB2.IsChecked && repeatableRB22.IsChecked && (dayOfMonthPicker.SelectedIndex < 0);
 
             if(isGoalNameEmpty || isGoalDescriptionEmpty || isWeeklyCheckedAndEntryFilled || isMonthlyCheckedAndEntryFilled)
             {
@@ -85,6 +83,7 @@ namespace HIPER
                 {
                     selectedGoal.Deadline = DateTime.Parse(goalDeadlineEntry.Date.ToString());
                 }
+                selectedGoal.LastUpdatedDate = DateTime.Now;
 
                 selectedGoal.Title = goalNameEntry.Text;
                 selectedGoal.Description = goalDescriptionEntry.Text;
@@ -106,17 +105,26 @@ namespace HIPER
 
         private async void deleteGoal_Clicked(System.Object sender, System.EventArgs e)
         {
-            await App.client.GetTable<GoalModel>().DeleteAsync(selectedGoal);
-            await DisplayAlert("Success", "Goal deleted", "Ok");
-            await Navigation.PopAsync();
 
+            bool delete = await DisplayAlert("Wait", "Are you sure you want to delete this goal?", "Yes", "No");
+            if (delete)
+            {
+                await App.client.GetTable<GoalModel>().DeleteAsync(selectedGoal);
+                await DisplayAlert("Sure", "Lets delete it", "Ok");
+                await Navigation.PopAsync();
+            }
         }
 
         private async void completeGoal_Clicked(System.Object sender, System.EventArgs e)
         {
             
             selectedGoal.Completed = true;
-            selectedGoal.Closed = true;
+            selectedGoal.LastUpdatedDate = DateTime.Now;
+            if (selectedGoal.RepeatType == 0)
+            {
+                selectedGoal.Closed = true;
+            }
+            selectedGoal.ClosedDate = DateTime.Now.Date;
 
             await App.client.GetTable<GoalModel>().UpdateAsync(selectedGoal);
             await DisplayAlert("Congratulations", "Goal completed", "Ok");
@@ -133,6 +141,8 @@ namespace HIPER
                 dayOfMonthPicker.IsEnabled = false;
                 weekdayPicker.IsEnabled = false;
                 goalDeadlineEntry.IsEnabled = true;
+                dayOfMonthPicker.SelectedIndex = -1;
+                weekdayPicker.SelectedIndex = -1;
             }
             else if (repeatableRB2.IsChecked)
             {
@@ -143,11 +153,13 @@ namespace HIPER
                 {
                     weekdayPicker.IsEnabled = true;
                     dayOfMonthPicker.IsEnabled = false;
+                    dayOfMonthPicker.SelectedIndex = -1;
                 }
                 else
                 {
                     weekdayPicker.IsEnabled = false;
                     dayOfMonthPicker.IsEnabled = true;
+                    weekdayPicker.SelectedIndex = -1;
                 }
             }
         }
